+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+* Improve Android DomeException error reporting
+* Add security key support for iOS
+* Add authenticatorAttachment, excludeCredentials and allowCredentials support for iOS
+
 ## 1.0.2
 
 * Update README